Make separate polybar modules for desktops and node flags/states
The number of node flags can change, meaning that label-mode's width is not
constant.  This leads to label-state (the desktop part) shifting around pretty
often, and that kind of annoys me.

I can't find a way to format non-tokens to a specific width to keep label-state
from shifting around, so I've split label-mode and label-state into separate
modules (bspwm-info and bspwm-desktops, respectively).  I left bspwm-desktops in
the center and placed bspwm-info on the right.

I also tried placing bspwm-info to the left of bspwm-desktops and countering the
shifting by placing a copy of bspwm-info to the right of bspwm-desktops, but
with spaces instead of icons.  However, this didn't work out either.